Improve the usage of conditional statement
Instead of single bracket (`[`), use double bracket (`[[`).

Not only double-bracket is an improvement to the single-bracket
(i.e. test command), but it improves the behavior of zsh's
automatic title change feature; with single brackets only,
one might end up in having only the 'MYOLDPWD' string in the
title of current zsh shell. This patch corrects the behavior.